Refactor and re-use getPriceAndStats
